Split tmux session overlays
